+++ src/main/java/com/lyz/ddedss_springboot/dto/resp/QueryAllUserListRespDto.java
@@ -23,7 +23,7 @@ public class QueryAllUserListRespDto implements Serializable {
     /**
      * 用户学号/工号
      */
-    private Integer userNumber;
+    private Long userNumber;
 
     /**
      * 用户名
